Renamed & re-styled card module
